// src/constants/types.ts
export interface Address {
  97?: string;
  56: string;
}

export interface SerializedToken {
  chainId: number;
  address: string;
  decimals: number;
  symbol?: string;
  name?: string;
  projectLink?: string;
  logoURI?: string;
}
// ...
interface FarmConfigBaseProps {
  pid: number;
  v1pid?: number;
  lpSymbol: string;
  lpAddresses: Address;
  multiplier?: string;
  isCommunity?: boolean;
  dual?: {
    rewardPerBlock: number;
    earnLabel: string;
    endBlock: number;
  };
}

export interface SerializedFarmConfig extends FarmConfigBaseProps {
  token: SerializedToken;
  quoteToken: SerializedToken;
}